fixed transactions

+++ schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE transactions (
     id SERIAL PRIMARY KEY,
     user_id INTEGER,
     coin TEXT,
-    value INTEGER,
+    value NUMERIC,
     history VARCHAR(255),
     created_at TIMESTAMP WITH TIME ZONE DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT fk_user
